LPS-55772 Remove circular dependencies on startup (HttpSolrServer)
See @a668bfe492e64ea65bcfa5a1fc0ef86ded2a2e01
