+++ webs/solr4-web/docroot/WEB-INF/src/META-INF/solr-spring.xml
@@ -71,7 +71,7 @@
 			</bean>
 		</property>
 	</bean>
-	<bean class="com.liferay.portal.search.solr.http.HttpClientFactoryImpl" destroy-method="shutdown" id="com.liferay.portal.search.solr.http.HttpClientFactory">
+	<bean class="com.liferay.portal.search.solr.server.HttpSolrServer" id="com.liferay.portal.search.solr.server.HttpSolrServer">
 		<property name="auth" value="${server.http.client.factory.auth}" />
 		<property name="httpClientFactories">
 			<map>
@@ -79,9 +79,6 @@
 				<entry key="CERT" value-ref="com.liferay.portal.search.solr.http.CertAuthPoolingHttpClientFactory" />
 			</map>
 		</property>
-	</bean>
-	<bean class="com.liferay.portal.search.solr.server.HttpSolrServer" id="com.liferay.portal.search.solr.server.HttpSolrServer">
-		<property name="httpClientFactory" ref="com.liferay.portal.search.solr.http.HttpClientFactory" />
 		<property name="url" value="${server.http.url}" />
 	</bean>
 	<bean class="com.liferay.portal.search.solr.server.LiveServerChecker" id="com.liferay.portal.search.solr.server.LiveServerChecker">
